+++ src/frontend/src/optimizer/plan_node/mod.rs
@@ -257,7 +257,6 @@ mod predicate_pushdown;
 pub use predicate_pushdown::*;
 
 pub mod generic;
-pub mod generic_derive;
 pub mod stream;
 pub mod stream_derive;
